fix newBoard modal
