+++ src/components/NewBoard/NewBoard.jsx
@@ -3,15 +3,12 @@ import { useDispatch } from 'react-redux';
 import { addBoard } from 'redux/boards/operations';
 import sprite from '../../images/sprite.svg';
 import Modal from '../Modal/Modal';
-import css from '../Sidebar/Sidebar.module.css';
 import styles from './NewBoard.module.css';
 
 import { backgrounds } from 'constants/backgrounds';
 import { icons } from 'constants/icons';
 
-const NewBoard = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const toggleModal = () => setIsModalOpen(state => !state);
+const NewBoard = ({ toggleModal }) => {
   const [icon, setIcon] = useState('icon-project');
   const [background, setBackground] = useState(null);
 
@@ -36,17 +33,7 @@ const NewBoard = () => {
   };
 
   return (
-    <div>
-      <button
-        onClick={toggleModal}
-        className={css.sidebarBoardButton}
-        type="button"
-      >
-        <svg className={css.sidebarBoardIcon}>
-          <use href={sprite + '#icon-plus'}></use>
-        </svg>
-      </button>
-      {isModalOpen && (
+
         <Modal onClose={toggleModal}>
           <form onSubmit={handleSubmit}>
             <h1 className={styles.title}>New Board</h1>
@@ -132,9 +119,7 @@ const NewBoard = () => {
             </button>
           </form>
         </Modal>
-      )}
-    </div>
-  );
+       );
 };
 
 export default NewBoard;
